move the types defined in parser package to the types package
