+++ src/lang-promql/parser/vector.test.ts
@@ -20,12 +20,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import { buildVectorMatching, VectorMatchCardinality } from './type';
+import { buildVectorMatching } from './vector';
 import { createEditorState } from '../../test/utils';
 import { walkThrough } from './path-finder';
 import { BinaryExpr, Expr } from 'lezer-promql';
 import chai from 'chai';
 import { syntaxTree } from '@codemirror/next/language';
+import { VectorMatchCardinality } from '../types/vector';
 
 describe('buildVectorMatching test', () => {
   const testCases = [
